add celery task to update master pipeline
python run.py nonbibToMasterPipeline # sends all nonbib records to master pipeline
as before, python run.py runRowViewPipeline # reads column files into Postgres
also removed nonbibUpdater and metricsUpdater, they copied data betweeen Postgres tables.
currently, the master pipeline holds the reference data rather than this repo so these updaters are no longer used.
