--- adsdata/metricsUpdater/app.py
@@ -1,12 +1,0 @@
-"""                                                                                                                     
-The main application object (it has to be loaded by any worker/script)                                                  
-in order to get a working configuration.                                                    
-"""
-from __future__ import absolute_import, unicode_literals
-from adsputils import ADSCelery
-
-
-class MetricsUpdaterCelery(ADSCelery):
-   
-    pass 
-
