+++ adsdata/row_view.py
@@ -414,7 +414,7 @@ class SqlSync:
        from {0}.Canonical natural left join {0}.Author \
        natural left join {0}.Refereed                 \
        natural left join {0}.Simbad natural left join {0}.Grants \
-       natural left join {0}.Citation                 \
+       natural left join {0}.Citation  natural left join {0}.Ned   \
        natural left join {0}.Relevance natural left join {0}.Reader \
        natural left join {0}.Download natural left join {0}.Reads   \
        natural left join {0}.Reference;'
