+++ requirements.txt
@@ -1,5 +1,6 @@
-git+https://github.com/adsabs/ADSPipelineUtils.git@master
 ConcurrentLogHandler==0.9.1
+git+https://github.com/adsabs/ADSPipelineMsg.git@master
+git+https://github.com/adsabs/ADSPipelineUtils.git@master
 requests==2.8.1
 python-Levenshtein==0.12.0
 sqlalchemy==1.0.8
